Deduplicate node:zlib compression into platform-node-shared
The one-shot zlib compression, streaming transforms, and algorithm
detection were copy-pasted across the Node, Bun, and Deno platforms.
Move them to a shared NodeHttpCompression module, which also removes
the Deno-local node:zlib/node:stream type shims.
