+++ packages/platform-bun/src/BunHttpPlatform.ts
@@ -9,82 +9,21 @@
  *
  * @since 4.0.0
  */
-import * as Effect from "effect/Effect"
+import * as NodeHttpCompression from "@effect/platform-node-shared/NodeHttpCompression"
+import type * as Effect from "effect/Effect"
 import type { FileSystem } from "effect/FileSystem"
 import * as Layer from "effect/Layer"
 import * as Etag from "effect/unstable/http/Etag"
-import * as HttpBody from "effect/unstable/http/HttpBody"
 import * as Platform from "effect/unstable/http/HttpPlatform"
 import * as Response from "effect/unstable/http/HttpServerResponse"
-import * as Zlib from "node:zlib"
 import * as BunFileSystem from "./BunFileSystem.ts"
-
-const brotliParams = (level: number | undefined, sizeHint: number): Zlib.BrotliOptions => {
-  const params: Record<number, number> = { [Zlib.constants.BROTLI_PARAM_SIZE_HINT]: sizeHint }
-  if (level !== undefined) {
-    params[Zlib.constants.BROTLI_PARAM_QUALITY] = level
-  }
-  return { params }
-}
-
-const zstdParams = (level: number | undefined): Zlib.ZstdOptions | undefined =>
-  level === undefined || level === 3 ? undefined : { params: { [Zlib.constants.ZSTD_c_compressionLevel]: level } }
-
-const compress = (
-  data: Uint8Array,
-  algorithm: Platform.CompressionAlgorithm,
-  options?: Platform.CompressionOptions | undefined
-): Effect.Effect<Uint8Array> =>
-  Effect.callback((resume) => {
-    const complete = (error: Error | null, result: Uint8Array) =>
-      resume(error === null ? Effect.succeed(result) : Effect.die(error))
-    switch (algorithm) {
-      case "gzip": {
-        Zlib.gzip(data, { level: options?.level }, complete)
-        break
-      }
-      case "deflate": {
-        Zlib.deflate(data, { level: options?.level }, complete)
-        break
-      }
-      case "br": {
-        Zlib.brotliCompress(data, brotliParams(options?.level, data.byteLength), complete)
-        break
-      }
-      case "zstd": {
-        const params = zstdParams(options?.level)
-        if (params === undefined) {
-          Zlib.zstdCompress(data, complete)
-        } else {
-          Zlib.zstdCompress(data, params, complete)
-        }
-        break
-      }
-    }
-  })
 
 // Bun's CompressionStream supports an extended format set covering brotli and
 // zstd
-const streamCompression = Platform.makeCompressionWeb({
+const compression = NodeHttpCompression.make(Platform.makeCompressionWeb({
   algorithms: ["gzip", "deflate", "br", "zstd"],
   transform: (algorithm) => Platform.compressionTransformWeb(algorithm === "br" ? "brotli" : algorithm)
-})
-
-const compression: Platform.Compression = {
-  algorithms: streamCompression.algorithms,
-  compressResponse(response, algorithm, options) {
-    const body = response.body
-    if (body._tag !== "Uint8Array") {
-      return streamCompression.compressResponse(response, algorithm, options)
-    }
-    return Effect.map(compress(body.body, algorithm, options), (result) =>
-      Response.setHeader(
-        Response.setBody(response, HttpBody.uint8Array(result, body.contentType)),
-        "content-length",
-        result.byteLength.toString()
-      ))
-  }
-}
+}))
 
 /**
  * @category constructors
