+++ packages/platform-node/src/NodeHttpPlatform.ts
@@ -8,6 +8,7 @@
  *
  * @since 4.0.0
  */
+import * as NodeHttpCompression from "@effect/platform-node-shared/NodeHttpCompression"
 import * as Effect from "effect/Effect"
 import { pipe } from "effect/Function"
 import * as Layer from "effect/Layer"
@@ -17,90 +18,10 @@ import * as HttpBody from "effect/unstable/http/HttpBody"
 import * as Platform from "effect/unstable/http/HttpPlatform"
 import * as ServerResponse from "effect/unstable/http/HttpServerResponse"
 import * as Fs from "node:fs"
-import type { Duplex } from "node:stream"
 import { Readable } from "node:stream"
-import * as Zlib from "node:zlib"
 import Mime from "./Mime.ts"
 import * as NodeFileSystem from "./NodeFileSystem.ts"
 import * as NodeStream from "./NodeStream.ts"
-
-const compressionAlgorithms: Array<Platform.CompressionAlgorithm> = ["gzip", "deflate", "br"]
-if (typeof Zlib.zstdCompress === "function") {
-  compressionAlgorithms.push("zstd")
-}
-
-const brotliParams = (level: number | undefined, sizeHint?: number): Zlib.BrotliOptions => {
-  const params: Record<number, number> = {}
-  if (level !== undefined) {
-    params[Zlib.constants.BROTLI_PARAM_QUALITY] = level
-  }
-  if (sizeHint !== undefined) {
-    params[Zlib.constants.BROTLI_PARAM_SIZE_HINT] = sizeHint
-  }
-  return { params }
-}
-
-const zstdParams = (level: number | undefined): Zlib.ZstdOptions | undefined =>
-  level === undefined || level === 3 ? undefined : { params: { [Zlib.constants.ZSTD_c_compressionLevel]: level } }
-
-const compress = (
-  data: Uint8Array,
-  algorithm: Platform.CompressionAlgorithm,
-  options?: Platform.CompressionOptions | undefined
-): Effect.Effect<Uint8Array> =>
-  Effect.callback((resume) => {
-    const complete = (error: Error | null, result: Uint8Array) =>
-      resume(error === null ? Effect.succeed(result) : Effect.die(error))
-    switch (algorithm) {
-      case "gzip": {
-        Zlib.gzip(data, { level: options?.level }, complete)
-        break
-      }
-      case "deflate": {
-        Zlib.deflate(data, { level: options?.level }, complete)
-        break
-      }
-      case "br": {
-        Zlib.brotliCompress(data, brotliParams(options?.level, data.byteLength), complete)
-        break
-      }
-      case "zstd": {
-        const params = zstdParams(options?.level)
-        if (params === undefined) {
-          Zlib.zstdCompress(data, complete)
-        } else {
-          Zlib.zstdCompress(data, params, complete)
-        }
-        break
-      }
-    }
-  })
-
-const compressTransform = (
-  algorithm: Platform.CompressionAlgorithm,
-  options?: Platform.CompressionOptions | undefined
-): Duplex => {
-  switch (algorithm) {
-    case "gzip": {
-      return Zlib.createGzip({ level: options?.level, flush: Zlib.constants.Z_SYNC_FLUSH })
-    }
-    case "deflate": {
-      return Zlib.createDeflate({ level: options?.level, flush: Zlib.constants.Z_SYNC_FLUSH })
-    }
-    case "br": {
-      return Zlib.createBrotliCompress({
-        ...brotliParams(options?.level),
-        flush: Zlib.constants.BROTLI_OPERATION_FLUSH
-      })
-    }
-    case "zstd": {
-      return Zlib.createZstdCompress({
-        ...zstdParams(options?.level),
-        flush: Zlib.constants.ZSTD_e_flush
-      })
-    }
-  }
-}
 
 // replaces the response body while keeping every other field, dropping the
 // now-stale Content-Length header
@@ -110,25 +31,17 @@ const compressedBody = (
 ): ServerResponse.HttpServerResponse =>
   ServerResponse.removeHeader(ServerResponse.setBody(response, body), "content-length")
 
-const compression: Platform.Compression = {
-  algorithms: new Set(compressionAlgorithms),
+const compression = NodeHttpCompression.make({
+  algorithms: NodeHttpCompression.algorithms,
   compressResponse(response, algorithm, options) {
     const body = response.body
     switch (body._tag) {
-      case "Uint8Array": {
-        return Effect.map(compress(body.body, algorithm, options), (result) =>
-          ServerResponse.setHeader(
-            ServerResponse.setBody(response, HttpBody.uint8Array(result, body.contentType)),
-            "content-length",
-            result.byteLength.toString()
-          ))
-      }
       case "Stream": {
         return Effect.succeed(compressedBody(
           response,
           HttpBody.stream(
             NodeStream.pipeThroughDuplex(body.stream, {
-              evaluate: () => compressTransform(algorithm, options)
+              evaluate: () => NodeHttpCompression.compressTransform(algorithm, options)
             }),
             body.contentType
           )
@@ -138,7 +51,7 @@ const compression: Platform.Compression = {
         const readable = body.body instanceof Readable
           ? body.body
           : Readable.fromWeb(new Response(body.body as BodyInit).body as any)
-        const transform = compressTransform(algorithm, options)
+        const transform = NodeHttpCompression.compressTransform(algorithm, options)
         readable.on("error", (cause) => transform.destroy(cause))
         transform.on("error", (cause) => readable.destroy(cause))
         transform.on("close", () => readable.destroy())
@@ -151,7 +64,7 @@ const compression: Platform.Compression = {
       }
     }
   }
-}
+})
 
 /**
  * Creates the Node `HttpPlatform`, serving file responses from Node readable
